pyproject.toml
[project]
name = "tagger-benchmark"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.14"
dependencies = [
    "openrouter>=0.1.1",
    "pillow>=12.0.0",
    "python-dotenv>=1.2.1",
    "timm>=1.0.22",
    "wdtagger>=0.16.0",
]
